Added renderpass function impls to source file
