+++ src/Engine.cpp
@@ -7,6 +7,7 @@
 #include "VoxelWorld.hpp"
 #include "VkZero/Internal/window_internal.hpp"
 #include "VkZero/Internal/renderpass_internal.hpp"
+#include "VkZero/Internal/raytracing_renderpass_internal.hpp"
 #include "shaders.hpp"
 #include <cstdio>
 #include <memory>
